refactor: delocalised.py
