+++ geometric/tests/test_diff.py
@@ -7,6 +7,7 @@ import numpy as np
 import json, os, shutil
 
 import geometric.coordinate_systems.cartesian
+import geometric.coordinate_systems.delocalised
 from . import addons
 import geometric
 import pytest
@@ -17,10 +18,10 @@ datad = addons.datad
 
 def test_diff_h2o2_h2o(localizer):
     M = geometric.molecule.Molecule(os.path.join(datad, 'h2o2_h2o.pdb'))
-    IC = geometric.internal.DelocalizedInternalCoordinates(M, build=True, connect=False, addcart=False)
+    IC = geometric.coordinate_systems.delocalised.DelocalizedInternalCoordinates(M, build=True, connect=False, addcart=False)
     IC1 = geometric.internal.PrimitiveInternalCoordinates(M, build=True, connect=False, addcart=False)
     IC2 = geometric.coordinate_systems.cartesian.CartesianCoordinates(M)
-    IC3 = geometric.internal.DelocalizedInternalCoordinates(M, build=True, connect=True, addcart=False)
+    IC3 = geometric.coordinate_systems.delocalised.DelocalizedInternalCoordinates(M, build=True, connect=True, addcart=False)
     assert IC1.repr_diff(IC) == "Primitive -> Delocalized"
     assert IC.repr_diff(IC1) == "Delocalized -> Primitive"
     assert IC.repr_diff(IC) == ""
